Fixed moving gump outside the screen.

+++ OrionUO/Network/Packets.cpp
@@ -411,7 +411,7 @@ CPacketUnicodeSpeechRequest::CPacketUnicodeSpeechRequest(const wchar_t *text, SP
 	//Sallos aka PlayUO algorithm
 	if (encoded)
 	{		
-		IFOR(i, 0, codeBytes.size())
+		IFOR(i, 0, (int)codeBytes.size())
 		{
 			WriteUInt8(codeBytes[i]);
 		}
